# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie Auditions API",
    "description": "API Documentation for Movie Auditions Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Inception"
          },
          "description": {
            "type": "string",
            "example": "A thief who steals corporate secrets..."
          },
          "genre": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "Sci-Fi"
            }
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "example": "2025-12-31"
          },
          "director": {
            "type": "string",
            "example": "Christopher Nolan"
          },
          "budget": {
            "type": "number",
            "example": 160000000
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "upcoming"],
            "example": "active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-04T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-04T12:00:00Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/admin/movies": {
      "get": {
        "tags": ["Admin Movies"],
        "summary": "Get list of movies (Admin)",
        "description": "Returns list of all movies for admin",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "description": "Filter by genre",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["active", "inactive", "upcoming"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Movies retrieved successfully."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": ["Admin Movies"],
        "summary": "Store a new movie (Admin)",
        "description": "Creates a new movie",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "genre", "end_date", "director", "status"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Inception"
                  },
                  "description": {
                    "type": "string",
                    "example": "A thief who steals corporate secrets..."
                  },
                  "genre": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "Sci-Fi"
                    }
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-12-31"
                  },
                  "director": {
                    "type": "string",
                    "example": "Christopher Nolan"
                  },
                  "budget": {
                    "type": "number",
                    "example": 160000000
                  },
                  "status": {
                    "type": "string",
                    "enum": ["active", "inactive", "upcoming"],
                    "example": "active"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role_type": {
                          "type": "string",
                          "example": "Lead Actor"
                        },
                        "gender": {
                          "type": "string",
                          "example": "Male"
                        },
                        "age_range": {
                          "type": "string",
                          "example": "25-35"
                        },
                        "dialogue_sample": {
                          "type": "string",
                          "example": "Sample dialogue..."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Movie"
                    },
                    "message": {
                      "type": "string",
                      "example": "Movie created successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/admin/movies/{id}": {
      "get": {
        "tags": ["Admin Movies"],
        "summary": "Get movie information (Admin)",
        "description": "Returns movie data with roles",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Movie"
                    },
                    "message": {
                      "type": "string",
                      "example": "Movie retrieved successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Movie not found"
          }
        }
      },
      "put": {
        "tags": ["Admin Movies"],
        "summary": "Update existing movie (Admin)",
        "description": "Updates a movie",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "genre", "end_date", "director", "status"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Inception"
                  },
                  "description": {
                    "type": "string",
                    "example": "A thief who steals corporate secrets..."
                  },
                  "genre": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "Sci-Fi"
                    }
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-12-31"
                  },
                  "director": {
                    "type": "string",
                    "example": "Christopher Nolan"
                  },
                  "budget": {
                    "type": "number",
                    "example": 160000000
                  },
                  "status": {
                    "type": "string",
                    "enum": ["active", "inactive", "upcoming"],
                    "example": "active"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "role_type": {
                          "type": "string",
                          "example": "Lead Actor"
                        },
                        "gender": {
                          "type": "string",
                          "example": "Male"
                        },
                        "age_range": {
                          "type": "string",
                          "example": "25-35"
                        },
                        "dialogue_sample": {
                          "type": "string",
                          "example": "Sample dialogue..."
                        },
                        "deleted": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Movie"
                    },
                    "message": {
                      "type": "string",
                      "example": "Movie updated successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Movie not found"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      },
      "delete": {
        "tags": ["Admin Movies"],
        "summary": "Delete movie (Admin)",
        "description": "Deletes a movie",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Movie deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Movie not found"
          }
        }
      }
    },
...
  }
}